modules/job-info: Check KVS for job id validity
In the job-info.list-id service, do not immediately return EINVAL
to the user if a job id is not in the job hash.  There is always the
chance that the ID is valid but has not yet been seen by the job-info
module.  Check the KVS to see if the ID is valid first.

Add new convenience code in idsync.[ch] to help track KVS lookups on
a list in the info_ctx called idsync_lookups.