libflux/response: export flux_response_derive()
Export an internal function that allows a response to be
crafted based on a request, but not sent right away.

We need to defer some responses in the job manager, and
using low level message functions to do it seems wrong.

Add coverage for bad args in the response unit test.