broker: fix possible deadlock in module unload
Problem: If a request or response is sent to a module in the middle
of shutdown (during a module remove), before the broker has fully
"removed" the module, but after the module has shut down its side
of the flux handle, then the main thread of the broker may block
trying to send the message and a deadlock occurs.

To fix, a module must make sure the broker has "muted" it so that
new messages are rejected *before* it processes unhandled requests.
To accomplish this synchronization the existing FLUX_MODSTATE_FINALIZING
keepalive message is used. During exit, module threads will synchronize
on this state by sending this keepalive message to the broker, which
"mutes" the module immediately. The broker then sends the same
keepalive message back to the module to ensure of synchronization,
and then the shutdown path proceeds as before.

Fixes #2709