cmd/flux-queue: add idle subcommand
Add flux queue idle [-t DURATION] [--quiet] to send a
job-manager.idle request and wait for its response with
optional timeout.  Upon return, the number of pending jobs
is printed.  If --quiet, suppress the count unless it is
nonzero.

This is like flux queue drain, except it waits for there to
be no pending alloc requests, and no jobs in RUN or CLEANUP state,
useful during shutdown.