modules/job-info: Wait for job id if not yet seen
If the job-info.list-id service has determined a job id is valid
but that it has not yet been seen or is still in state FLUX_JOB_NEW,
wait for that job id to be ready for a response later.

We accomplish this by storing waiting job ids on a idsync_waits
hash in the info_ctx.  When a job transitions to the FLUX_JOB_DEPEND
state, check the hash to see if any lookups via job-info.list-id
need to be responded to.  Respond if necessary.

Fixes #2592