broker: remove ill-advised flux_log in module_sendmsg
Problem: In commit 396bef35a05a9816598fcd44f86fb05d0dfadef0
a call to flux_log() was added in module_sendmsg() to note
when a message sent to a module is "muted" due to the FINALIZING
state. However, this flux_log() uses the module's side of the
flux_t handle from the broker, and this is not safe! Much
sadness will be caused by leaving this in.

Since module_sendmsg() (currently) doesn't have access to the
broker's handle, just drop the debug message for now.