job-ingest: cover all cases of unexpected worker exit
Problem: Handling of unexpected worker exit, i.e. processes that exit
before worker_stop() is called, was added in 931d9e745. However, the
code only handles workers that exit with non-zero exit status, leaving
workers that die due to a signal or exit unexpectedly with exit code 0.

Move the check for w->p == p to the end of the function after other
cleanup activities have completed.

Skip the unnecessary addition of p to w->trash, since zlist_remove
is a no-op if `p` is not in the list.