job-ingest: fix handling of unexpected worker exit
Problem: When handling unexpected EOF in worker_output_cb(), all
pending requests in the worker queue are failed. However, this
doesn't cover the case where a worker process A exits normally and,
before its EOF is received, a new worker B is started with a queued
request. When the EOF for A is received, the pending request for
B is aborted.

Handle this case by checking that w->p == p before assuming that the
exit of a process is unexpected.

Fixes #2718