broker/shutdown: change protocol to avoid rc3 timeout
Rework the broker shutdown protocol so that the "grace timeout"
does not span rc3 execution.  It starts after rc3 finishes so
that there is an unlimited amount of time available to sync
the instance data to disk.

The new protocol is documented at the top of broker/shutdown.c.
Rank 0 sends a shutdown event, then each broker waits for
its descendants to disconnect in before disconnecting from its
parent and exiting.  The TBON thus is torn down starting with the
leaves and leading to the root.  It ensure that routing (interior)
brokers stay up long enough to route the shutdown message to all
the leaves.

The broker no longer needs to run out the shutdown grace period
every time, so this fixes #1006.

There are a couple other changes here:

1) if rc3 fails, the instance fails

This seems appropriate given that rc3 may be writing out critical
data from the content cache or terminating jobs in an orderly fashion.
If it fails we need to know. Updated one of the tests in
t0014-runlevel.t to reflect this change

2) signal handling

This area needs work.  Before, the shutdown code would be invoked,
with an exit code and reason carried in the shutdown event so that
rank 0 could report it.  This will leave the instance in a completely
messed up state if it's running work.  For now just log something
to stderr and call _exit(1), to avoid creating the impression that
something reasonable is happening.

Drop unit test of old shutdown code.