testsuite: don't tune shutdown grace period
Problem: workarounds for #1006 (always running out the
shutdown grace period) are no longer required in the test
suite to run back to back instances in a reasonable amount
of time.  Also, if the grace period actually expires,
it now causes the instance to exit with a nonzero exit code.

Drop workarounds.