flux-jobs: refactor output format code into a class
Add a convenience class OutputFormat to flux-jobs, used to store
and manipulate the default or user-provided output format.

The OutputFormat class uses string.Formatter to parse the provided
format, which is more likely to provide sensible errors early (just
after argument parsing), rather than later when the first job entry
or header is printed.

Additionally, since string.Formatter parses the format into tokens,
we can more easily sanitize the format for use in printing the
header.

Since the format is kept internally in the OutputFormat object,
this class now provides `header` and `format(job)` methods that
generate the output lines for the utility.

To avoid unecessarily fetching jobs when there is an error parsing
the format, the OutputFormat is constructed before job gathering.