modules/content-sqlite: drop reload-to-cache mode
Problem: content-sqlite copies the sqlite db back
to memory when it unloads, unless it receives the
instance 'shutdown' event first.  When shutdown
is rearchitected and the event is and sent after
rc3 completes, this code will misbehave.

This was intended to allow someone to unload
content-sqlite and load a different content back-end
on a live system, right after instance startup.
It was not an elegant solution since it will fill
memory and eventually fail if attempted later on,
when stored content exceeds available memory.

Since this will no longer work and is not currently
used anyway, drop the reload-to-cache code and
associated tests.  Now content-sqlite does not
require the shutdown event "hint" to behave as
it did before in rc3.