flux-jobs: add JobInfo class and refactor
Add a JobInfo class as a container object for data returned from
job-info.list RPC. The class implements a getattr interface for both
the raw data returned in the RPC as well as all derived properties.

The .format method is switched to a getattr interface, so that a
JobInfo object can be passed as the sole argument to format.  This is
hidden from the user by substituing `0.` in front of attributes passed
to `-o, --format` that do not already have the required prefix.

This has the benefit of delaying creation of job attributes only
when they are requested as a result of being in the output format.
This will allow more efficient addition of alternate attrs in the
future.