flux-module: make unload a synonym for remove
Problem: Weaker minds continually want to type "unload" as the
antonym to "load" in the flux-module command, rather than "remove".
While we should be careful about catering wholesale to the indigent
Flux admin, the occasional mollification may engender good will
in the course of time.

Make `unload` a synonym for `remove` in flux-module, but don't tell
anyone about it so it is a pleasant surprise.