build: don't set -Werror=missing-field-initializers for clang-3.x
Problem: Versions of the clang compiler < 6.x throw an error when
-Werror=missing-field-initializers is set and static intializers
such as

  static struct name = {0};

are used.

This usage is quite convenient, and furthermore appears to be perfectly
valid C99, evidenced by the fact that all GCC compilers and Clang >=
6.x seem to accept this usage without warning.

Work around the issue and allow flux-core to be built with clang
{3,4,5}.x by disabling -Werror=missing-field-initializers for these
compiler versions.

This requires pulling in a new macro from the autoconf-archive:
AX_COMPILER_VERSION, and using this to test for clang < 4.0.