broker/overlay: set ROUTER_MANDATORY zmq sock opt on TBON
Problem: downstream TBON messages are silently dropped if
peer disconnects.

Ideally we'd like to at least get a log message if a message
cannot be routed to a TBON peer.  The default behavior of
a zeromq ROUTER socket is to drop messages that specify
an unknown peer in their route stack.

Add the ROUTER_MANDATORY socket option to the TBON ROUTER
socket.  This changes the behavior of the socket to return
an EHOSTUNREACH error instead of silently dropping an
unroutable messsage.

Fixes #2751