bindings/python: add JobWaitFuture and SubmitFuture
Subclass Future in the job.submit_async() and job.wait_async()
functions to customize the getter function available when the
Futures returned from these functions are fulfilled.

The old functions are left intact since they have been around
awhile and may have users in the wild.

Subclassing with new methods would seem to be a more Pythonic
approach in general, but is also more user-friendly and the result
is more readable.