bindings/python: add JobListRPC class to job module
Subclass RPC with JobListRPC and return an object of this class
from job.job_list(). This class exports a getter to return the
jobs array from the RPC object.

The old job.job_list_get() function is retained for now since it
is used in flux-jobs.py.

Subclassing RPC is probably more Pythonic, as well as making the
resulting code more readable.