broker: fix misrouted response messages
Problem: response messages addressed to a service belonging
to a comms module that is no longer loaded receive EHOSTUNREACH
errors now that ROUTER_MANDATORY is enabled.

broker_repsonse_sendmsg() directs a response message downstream
on the TBON if it fails to deliver to a comms module. Before
ROUTER_MANDATORY, this was a silent drop.  Now it's an
EHOSTUNREACH error, which is inappropriate.

Since we can easily tell if a route hop is a TBON broker rank,
use this information to make an informed routing decision for
routing up or down the TBON.  If a message cannot be routed
to a TBON destination, return EHOSTUNREACH.  If a message cannot
be routed to a module destination, return ENOSYS.

Also: clean up and refactor broker_response_sendmsg().
It was confusing.