modules/job-ingest: add shutdown timer
Problem: if validators don't stop when their stdin is closed,
job-ingest unload will hang.

As an added safeguard, arm a timer while waiting for validators
to stop.  If the timer expires, stop the reactor and fall
through to the previous mode of signaling them and exiting
while the broker handles cleanup.