broker/module: don't pthread_cancel module threads
Problem: module pthread_cancel() can lead to libzmq
broker hangs in zsys_shutdown()

zeromq sockets are not currently thread safe.  If a module
thread is canceled while it is inside a zeromq socket
call - see pthreads(7) cancellation points - the module's
PAIR socket may be left in an inconsistent state, and efforts
to clean it up from the broker thread may get into trouble.

Broker hangs occasionally occured when connector-local was
canceled on a fast cycling instance.  The observations were:
- the "dangling 'PAIR' socket" message appears during shutdown
- the zsys_shutdown() atexit handler hangs in zmq_ctx_term()

It seems that neither the prophylactic flux_close() following
the cancel, nor the zsys_shutdown() cleanup of dangling sockets
(both called from the main broker thread) actually affected a
close of the module's socket.  zmq_ctx_term(3) is documented to
block until all sockets are closed.

Remove the pthread_cancel() code.  Since it assumed only modules
in FLUX_MODSTATE_INIT should be canceled, and all modules are
now given enough time to complete initialization, enter the
reactor, and handle the 'shutdown' request naturally before the
broker reactor loop exits, this code is unlikely to ever be
invoked anyway.

If we find the broker getting stuck in pthread_join() again
because modules are blocked for other reasons, we will want
to add code that logs some debugging info about the stuck module
so that it can be debugged.  But canceling is contraindicated.