shell: mpir: increase buffer size for executable paths
Problem: The executable path is encoded in the shell procdesc
service by using the nodelist code, since this will efficiently
encode ranges of duplicates. However, the nodelist encoding has
a limit of MAXHOSTNAMELEN, which is much shorter than the maximum
path length of an executable, resulting in errors such as:

 shell-661005533184.proctable: Argument list too long

for moderately long paths.

Increase the name buffer size to 4K to accomodate long executable
paths.

Fixes #2757