broker: unload connector-local with reactor running
Problem: unloading the connector-local module synchronously,
without the broker reactor running requires specialized
code and is prone to deadlock.

The connector-local module is loaded and unloaded before
and after the reactor runs.  It is the only module hardwired
in the broker, because it is needed so that configurable
modules can be loaded from rc1 and unloaded in rc3.

Now that the shutdown grace period is not obligatory, a fast
cycling instance is more likely to attempt to unload connector-local
before it has entered the reactor.  This triggers a code path that
pthread_cancel()s the module thread, which is prone to deadlock.
That was added in part because connector-local makes a getattr RPC to
the broker in its initialization that cannot be responded to if the
broker reactor is not running.

Since we are now performing TBON teardown with the reactor running,
we can piggyback the unload of connector-local on that code and
thus avoid the specialized synchronous unload.