testsuite: drop racy test
Problem: test that attempts to force instance to time out
with very short grace period fails in travis.

It may not be possible to guarantee timer goes off faster
than handshake can be completed, since the final disconnect
event may already be in the socket queue when the timer is set.

Drop the test.