testsuite: only run t3000-mpi-basic.t with FLUX_TEST_MPI
Problem: The MPI tests in t3000-mpi-basic.t are brittle in the face
of the numerous possible distributions and versions of MPI that
might be found on an installed system. Furthermore, the default
MPI may not be tolerant of the way in which the test runs (i.e. the
MPI might not be tolerant of being oversubscribed)

To limit new user confusion and unnecessary failures of 'make check',
make these MPI tests opt-in with a FLUX_TEST_MPI environment variable.

Fixes #2760