cmd/flux-job: add flux job status
Add heavyweight `flux job status` to return the job completion
status of one or more jobs. This command is much more heavyweight
than flux-job wait, but will work for non-waitable jobs.

As a side-effect of the way the command works, by watching the job
eventlog for the "finish" event, this command may be used on running
jobs and will block until all jobids complete.