rc1: increase local connector timeout
Problem: when running under valgrind, the default
timeout waiting for flux_open() to succeed on FLUX_URI
is a bit too short, after timing changes in pr #2733.

This was noted in the discussion for #2761, when a
test run under valgrind unexpectedly failed on master.

Increase the timeout in rc1 only.