rc1: do some work in parallel
Problem: rc1 is annoyingly slow

Run some of the work in parallel.  The work selected to run
in parallel is exactly as proposed and tested in #2771.

Use the shell function wait_check() to wait on a list of pids
sequentially, to ensure that if any commands fail, the failure
is captured and rc3 exits due to being run with bash -e.

Fixes #2771