modules/kvs: Update namespace event subscription
On rank 0, we have to subscribe to all events on all KVS namespaces,
as rank 0 must have knowledge of all KVS namespaces.  On the other
hand, workers only need to know about the KVS namespaces that
callers are using on that rank.

To improve KVS namespace create performance, subscribe to all KVS
namespaces on rank 0 once on initialization, instead of subscribing
to each KVS namespace as it is created.  On workers, subscribe
to only the KVS namespaces that are necessary.

Fixes #2727