broker: drop persist-directory, persist-filesystem
Drop code that validates a 'persist-directory' attribute,
if set, or that sets 'persist-directory' to a unique
directory in 'persist-filesystem', if set.

Drop code that sets the broker 'log-filename' attribute
to a location in 'persist-directory', if the former is
unset and the latter is set.

Drop code from rc3 that attempts to write out a
KVS snapshot if 'persist-directory', if set.

Drop relevant sharness tests from t0001-basic.t.
Drop all tests from t2010-kvs-snapshot-restore.t for now.
Drop these attributes from flux-broker-attributes(7) man page.