content-sqlite: handle content.backing-path attr
Drop the code that manipulates the location and persistence
of the sqlite db file based on the 'persist-directory'
attribute, and replace it with the following:

If 'content.backing-path' attribute is set on entry,
use its value as the sqlite db file.  sqlite creates
this file if it doesn't exist.  On exit, it is preserved.

If 'content.backing-path' is not set on entry, use
'${rundir}/content.sqlite' as the sqlite db file.
sqlite creates this file if it doesn't exist(*).
On exit it is removed.  Set the 'content.backing-path'
attribte so the location of this file.

Add a 'flux content flush' command to rc3 just before
content-sqlite unload, to ensure that any dirty cache
entries are flushed to the backing store before content-sqlite
is unloaded.

_______
(*) This file is not expected to exist initially given that
rundir is a temporary directory, but it _could_ exit if the
content-sqlite module is reloaded.  Reusing its content
is appropriate in that case.