modules/kvs: save/restore checkpoint for primary ns
On rank 0 startup, try to get a 'kvs-primary' key from the
kvs-checkpoint service.  If the service is available and the
key is found, use it as the blobref for the initial KVS root.
If unavailable or not found, start with an empty directory as
before.

On rank 0 unload, try to put 'kvs-primary' to the kvs-checkpoint
service.  If the service is unavailable, don't treat this as
an error, since the instance may be purposefully running without one.