runlevel: add runlevel_abort() on signal
Problem: broker signal handler calls _exit(1), which bypasses
clean shutdown.  Systemd terminates the broker with SIGTERM
when flux unit is stopped, so clean shutdown is not possible
under systemd.

Add runlevel_abort() to terminate the current runlevel on
receipt of a signal on rank 0.  If there is a subprocess,
it is sent a SIGTERM.

Fixes #2789