broker: add broker.rc2_none attribute
Problem: the system instance must run an initial
program like 'sleep inf' to keep rc2 running until
shutdown.

Add a broker attribute 'broker.rc2_none', which if set,
ensures that rc2 is not configured with a subprocess.

Change the runlevel logic so if rc2 is not configured,
it is not immediately considered finished like rc1 and
rc3.  rc2 can be terminated (without error) by the broker
with a call to runlevel_abort().