python: use compiler to preprocess headers
This patch removes the need for make_binding, providing explicit build
scripts for core and security and using the C preprocessor to do the
preprocessing of headers in combination with a set of headers from the
pycparser library.  All the old python-based string munging is now gone,
and in principle some functions should be callable now that weren't
workable in CFFI in the original.  This also installs the generated
header files for use with a separate setup.py installation to be worked
out later.