scripts: make sure pylint runs on cmds as well as bindings
We have several times now created a class or function in a command,
found it useful elsewhere, lifted it up into the bindings, and then
gotten many pylint errors.  This was because pylint was only running on
the bindings, not on the cmds.

Setup pylint to run on python commands within src/cmd via a `pylint`
script made in the same vein as `format`. As a consequence of this move,
also move the pylint-related code in src/bindings/python/Makefile.am
into the top-level Makefile.am.

Since pylint assumes that the files being linted are importable files,
many of the errors do not make sense in the context of a command (i.e.,
`no-member`, `missing-docstring`). Pylint also enforces certain naming
constraints for the same reason.  Override those when running against
the commands (but leave them enabled on the bindings).