systemd: add TimeoutStopSec=90
Problem: a site may configure systemd with a short
DefaultTimeoutStopSec in /etc/systemd/system.conf, which may
not allow sufficient time for the system instance rank 0 broker
to shut down cleanly after receiving the first SIGTERM.

Explicitly set TimeoutStopSec=90 in the unit file to allow up
to 90 seconds for shutdown.  This is the default value for
DefaultTimeoutStopSec.