job-exec: run flux-imp if job->userid != current uid
If the job-exec module has been configured with a path to the IMP,
(i.e. either the 'exec.imp' config parameter or cmdline paramter
imp=  has been set), then attempt to run any job for a userid
other than the instance owner under the IMP.

If the job userid does not match the owner userid, then a 'multiuser'
flag is set and in addition to the jobspec and R, the signed J is
fetched from the kvs.

Then instead of running the job shell directly, the exec implementaiton
runs the configured IMP executable and sends it the requisite input.

If the IMP is able to verify that J is valid, it will invoke the job
shell as the user that signed J.