broker/ping: fix final hop of route string
Problem: route string constructed for ping response tacks
on the broker rank as a standin for the destination uuid,
but that is only correct for a broker not a module.

Add an argument to ping_initialize() to allow either the
stringified rank (broker) or a uuid (module) to be passed in.
Use this argument to construct the ping response.

Update broker.c and modservice.c.

Fixes #2810