broker: shutdown cleanly on receipt of signal
Problem: SIGTERM is fatal to the broker unless it is
received exactly once on rank 0.

Change broker signal handler to ignore signals when:
- rank > 0, or
- rank == 0 and runlevel == 3

On rank 0, in runlevel 2: if a signal is received multiple
times call runlevel_abort() each time.  When there is a subprocess
(initial program), this delivers SIGTERM to the subprocess on each
call, in case we have a stubborn one.  When there is no subprocess,
the runlevel would have immediately transitioned to runlevel 3 on
the first signal, so that case is unaffected.

This should behave reasonably for systemd stop of a size=1 system
instance, and for cancellation of any size flux instance launched
by Flux or another resource manager.

Add logging of each signal at LOG_INFO level.

Fixes #2812