fluid: add ability to restart generator
Problem: there is no way to save/restore fluid generator
state across a module reload or instance restart.

Add a timestamp argument to fluid_init() which specifies
a starting timestamp instead of assuming zero.

Add two functions for retrieving timestamps that can be used
to bootstrap a generator:
1) fluid_get_timestamp() extracts timestamp from a FLUID.
It may be useful to save the last allocated FLUID, then use
it to restart a generator from the FLUID's timestamp + 1
2) fluid_save_timestamp() obtains an up to date timestamp
from a generator.  The timestamp can be used to bootstrap a
(possibly late-joining) generator peer.

Also [cleanup]: to handle the case of a user requesting more
than 1024 FLUIDs from one generator ID within a 1ms time quanta,
the sleep + recursion in fluid_generate() is replaced with a
busy-wait.  Recursion in an unlikely code path seems like it is
asking for trouble.

Update unit test and job-ingest module.

Fixes #1545