job-exec: allow default shell path to be configured
Problem: The only method to override the builtin value for the job
shell path used by the job-exec module is a broker attribute. This
is inconvenient as it requires editing an rc file to "stick".

Allow job-exec execution "implementations" to configure themselves
during module initialization with a new config callback, and use this
callback in the normal 'exec' implementation to read a 'job-shell'
value from the [exec] config table. This overrides the any builtin
default shell path.

Finally, support overriding both the default and/or configured
job-shell path via a commandline job-shell= parameter.