job-exec: allow kill-timeout to be configured
Problem: the only method to override the builtin default exec
kill timeout is via a broker attribute. This is inconvenient since
it requires editing an rc file to "stick"..

Use the flux_conf_t facility to attempt to load configuration at
module initialization. Allow a key 'kill-timeout' to be set in a
new "[exec]" table in the TOML config to override the builtin
default of 5s. Furthermore, allow a module cmdline parameter
'kill-timeout=' to be specified which can override *both* the
builtin and configured defaults.

Both the configuration value and cmdline option are specified in FSD.