job-ingest: init fluid generator properly
Problem: The fluid generator in the job-ingest module is always
initialized with timestamp = 0, which means job-ingest can
generate duplicate jobids if the module is reloaded or the
instance is restarted.

On rank 0, ask job-manager for max_jobid (which job-manager ensures
persists across a restart).  Extract the timestamp from this jobid
and initialize the fluid generator with timestamp + 1.

On rank > 0, ask upstream job-ingest for current timestamp
and initialize the fluid generator with it.

Fixes #2816