job-exec: fix potential memory leak
Problem: In jobinfo_start_init() several kvs lookup futures
are pushed onto a composite wait_all future, but if any of
the flux_future_push() calls fail, then a future is leaked.

In the exit error path, be sure to free f_kvs if it is non-NULL
to avoid a leak.