sharness: add a per-test global timeout option
After recent frustration with a variety of hangs, this provides a new
environment variable `FLUX_TEST_TIMEOUT` that provides each individual
top-level sharness test (test_expect_success or similar) with an
individual timeout of that value in seconds.  After that many seconds
the sharness script receives a signal, prints an error, kills the
current command, and proceeds to subsequent tests.  The mechanism for
this is a little bit awkward because of the way that sharness uses eval
to execute tests, but has proven reliable in my tests so far.