configure: increase minimum pylint version to 1.8.4
Pylint 1.7.0 was the first to support py3.6, but the pip
package is broken (has no upper-bound on the astroid dependency) so
1.7.6 is the earliest pip-installable pylint that supports py3.6.

Unfortunately, 1.7.6 sees ghosts and claims there is a cyclic-import
that doesn't exist.  The 1.8+ series does not see ghost, but the earlier
versions have the same pip problem.  Use the latest patched version of
the 1.8 series, which is 1.8.4.