testsuite: add job personality to t26xx-job-shell tests
Problem: Some job shell tests were not run with the 'job'
test_under_flux personality, and thus were not getting the fake
resource.hwloc.by_rank that is automatically set by the rc1-job
script. This could cause the tests to fail on systems with only a
single core, or when bound to only one core.

Apply the 'job' personality to these tests to fix these potential
failures.