python: default config_pattern to None in SecurityContext()
Problem: Currently, in order to initialize a SecurityContext object
using the default config pattern (the common case), requires calling
SecurityContext(None). This is not only a bit silly, but is also
unintuitive and may push users to try to hard code a real pattern
which will be prone to errors.

Set the default pattern to None so users can call SecurityContex()
with no args.