tests: rework run_timeout to avoid alarm overload
The job-info rework with run_timeout was not functioning for any test
that also used the --timeout flag on the wait-event command.  The only
reason I can think of for this is overloading the meaning of SIGALRM
somewhere, and my perl is sufficiently rusty and the dep sufficiently
undocumented, so I decided to use python to make a more flexible
run_timeout.  This version uses the builtin timeout functionality of the
subprocess module for the timeout, propagates return codes, and to
maintain compatibility on timeout returns the alarm return code of 142.
It is slightly slower, but substantially more likely to be and remain
correct in the presence of signal handling in the executable being run.