python: mypy all the things
Moved .mypy.ini to the top, which allows `mypy` run at the root to
inherit all configs.  It now runs a type check on nearly all of our
python code.  The exceptions are errors are ignored in flux-jobs.py and
one test temporarily while a solution is found to generate type stubs
for at least flux.constants.  Without that, they cannot type check.