run_timeout: extend with env and kill_after args
This version is significantly more powerful, and is almost a complete
portable implementation of /usr/bin/timeout (only main difference is it
preserves status by default, because our tests expect that, and kills
after timeout by default after one second).  It additionally has the
ability to set environment variables on the command it is going to run,
this is the workaround for unsetting then re-setting LD_PRELOAD for the
faketime tests without breaking everything.