tests: fix all tests that depend on exit 142
All of these expected 142 from a job signaled by run_timeout.  I decided
to preserve that and add the `-s ALRM` argument to run_timeout to
preserve the behavior, in future since the default is KILL it may be
worth expecting the exit code for KILL, but some of these are correct
cleanup tests so I didn't want to break the intent of the test.