libflux: don't parse config file before flux_open()
Problem: flux_open() attempts to parse TOML config if
FLUX_URI is not set, but we would like to use broker options
to select the config path, and the selected path cannot be
known before flux_open().

This was implemented for a specific use case: a login
node that doesn't run a flux broker.

Drop this code from flux_open().

We may be able to address the use case by simpling setting
FLUX_URI to ssh://... in the default user environment.

Fixes #2856.