broker: add --config-path option, share w/ modules
Create broker config object using (descending precedence):
  1) Specify TOML directory with --config-path=DIR
  2) Specify TOML directory with FLUX_CONF_DIR env var
  3) empty config object

Cache the config object in the broker's flux_t handle
using flux_set_conf() so that flux_get_conf() may be used
without error checking in the rest of the broker.

As part of module creation, pre-set the module flux_t handle
with a copy of the config object so flux_get_conf() works
in modules as well, without requiring the configuration to be
parsed there.

Isolate the broker config code in 'brokercfg.c' as it
will need to expand in the future to handle config reload,
and a mechanism for distributing updates.

Set the 'config.path' broker attribute to reflect the
configured path.  To access the config outside of the broker and
its modules, for example from a flux utility, use this method:

  const char *path = flux_attr_get (h, "config.path");
  flux_conf_t *conf;
  flux_conf_error_t error;

  if (!(conf = flux_conf_parse (path, &error)))
	/* handle error */