job-exec: use new flux_get_conf() prototype
Problem: Now that the config object is parsed only in
the broker, and modules are constructed with a copy of
it attached to their flux_t handle, flux_get_conf() has a
new prototype and need not be checked for errors in modules.

Adjust flux_get_conf() call sites for new prototype.
Drop bad TOML test from t2403-job-exec-conf.t.