broker: eliminate boot.method attribute
Problem: the broker selects its TBON wire-up mechanism
based on the boot.method attribute, but now that a user
instance doesn't have to read the same config as the
system instance, this attribute is redundant with the
config itself.

Let the config file select the bootstrap method.  If the
[bootstrap] stanza is present, then use its hosts table
to bootstrap, otherwise fall back to PMI.