connector-local: use config not userdb
Problem: sites will most likely prefer to configure their Flux
access policy using config files, but the only way to do that
now is with the "flux user ..." command.  Also, the connector-local
needs to make an RPC to userdb on rank 0 to authenticate guests.

Use static configuration to configure access policy.
The connector-local module can read this once at startup and apply
it to every connection, without the overhead of an RPC.

If there is no configuration, then only the instance owner may access.
To allow guests, configure:

[access]
allow-guest-user = true

N.B. this is a strong use case for config reload, but deferring
that to another PR.

Fixes #2845
Fixes #2844