broker:  add config.reload handler
Add a handler for the config.reload RPC.  First, TOML
files at the current config path are re-parsed.  Then
each loaded module is sent a '<name>.config-reload' RPC
containing the new config object.

If TOML parsing fails, a human readable error is returned
in the RPC and the config is unchanged.  If TOML parsing
succeeds, and one or more module config-reload RPCs fail,
human readable error(s) are returned in the RPC, but
any config changes that were accepted remain in place.

There is not currently a hook for "broker services" to
be notified when the config changes, however this can
be added when there is a use case.  (Perhaps a stack of
callbacks can be pushed in the 'brokerfcfg' "class").