libflux: add flux_conf_reload_decode()
Problem: there is no way to translate the JSON config
object from the '<name>.config-reload' RPC into a new
flux_conf_t.

Add flux_conf_reload_decode() which creates a flux_conf_t
from the request payload, and provides access to it.
The flux_conf_t destruction is tied to the request message,
so it must be flux_conf_incref()ed in order to be retained.