connector-local: handle config-reload
Add a handler for the 'connector-local.config-reload' RPC.

If the [access] table contains errors, keep the existing
configuration and return an error to the broker.

Log configured values so they can be checked with dmesg.

Also, tighten up the JSON parsing format string so that any
extra keys in the [access] table trigger an error, for example,
if a user misspelled one of the valid keys.