shell: initial pty support for jobs
Add an initial implementation of psuedo-tty support for jobs.

Add a plugin that starts a "terminus" server in the shell.init callback
in each job shell, allowing a client to open an interactive shell to
any job shell within a job.

Additionally, if the "pty" shell option is used, open a session for
task 0 *only*, and run the task attached to this session, allowing
flux-job attach or other client tool to optionally attach to this
pty at <shell-service>.terminus.0 on the shell leader rank.

Add context to the shell.init event that indicates where the pty server
is listening. This allows a utility to connect to this pty at service
"<uid>-shell-<jobid>.pty.0" on rank "leader_rank".

Fixes #282