flux-job: attach to pty if job has one
If the shell.init event notes that the job has spawned a pty
on rank 0, connect as a pty client to the job's pty service by
default, instead of processing normal stdio.

This has several side effects, including:

 * flux-job attach will exit when the remote pty is terminated,
   not necessarily after all tasks have exited.
 * flux-job attach will exit with non-zero status only if the pty
   is disconnected with non-zero status, instead of exiting with
   the status reported in the exec eventlog.
 * flux-job attach can only be "disconnected" from the job with
   the pty detach character (`^\` by default) not `^C^Z`.

It is assumed that the above is the least surprising behavior for
jobs that have requested the pty option. In the future, these
behaviors could be tuned with further options to flux-job attach.