libutil: llog.h: add library logging wrapper header
Problem: Flux internal libraries that are used in multiple contexts
do not have a good strategy for generating useful log messages.

The llog interface, borrowed from libtsm[1], is a simple header-only
solution that allows a library or module to implement logging, yet
export the actual loggin implementation to the user of the lib/module.

In this way, a Flux convenience library can continue to implement
logging, with the logging destination chosen by the component
using the library, e.g. when used in a utility, log messages can
be sent to stderr, when used in a broker module, flux_log(3) can
be used, and if used in the job shell, messages can be sent to the
shell logging facility.

See the llog.h header for more information.

[1] https://www.freedesktop.org/wiki/Software/kmscon/libtsm/