libtestutil: ensure requests have a "sender"
Problem: Because the libtestutil test_server implementation uses
directly connected zeromq sockets with no routing, a "sender" identifier
cannot be obtained for requests by querying the first route on the msg
with flux_msg_get_route_first(). This means that services relying on
obtaining the sender identity cannot use the test_server implementation.

As a special case, push a "fake" first route onto requests in the
test connector, and pop it back off for responses. This mocks a routed
message in the test server, and allows flux_msg_get_route_first() to
work as expected.